Address All-XY review: rename to 20, DRY sequences, and score outcomes
