+++ qualibration_graphs/superconducting/calibration_utils/all_xy/plotting.py
@@ -10,32 +10,9 @@ from matplotlib.figure import Figure
 from qualibration_libs.plotting import QubitGrid, grid_iter
 from quam_builder.architecture.superconducting.qubit import AnyTransmon
 
-# Labels for the 21 All-XY sequences (for x-axis). Order matches ALL_XY_SEQUENCES in the node.
-ALL_XY_LABELS = [
-    "I,I",
-    "x180,x180",
-    "y180,y180",
-    "x180,y180",
-    "y180,x180",
-    "x90,I",
-    "y90,I",
-    "x90,y90",
-    "y90,x90",
-    "x90,y180",
-    "y90,x180",
-    "x180,y90",
-    "y180,x90",
-    "x90,x180",
-    "x180,x90",
-    "y90,y180",
-    "y180,y90",
-    "x180,I",
-    "y180,I",
-    "x90,x90",
-    "y90,y90",
-]
-# Ideal outcomes: 5 excited (max), 12 superposition (mean), 4 ground (min)
-N_EXCITED, N_SUPERPOSITION, N_GROUND = 4, 12, 5
+from .sequences import ALL_XY_LABELS, N_EXCITED, N_GROUND, N_SUPERPOSITION
+
+# Ideal outcomes along sequence index: ground, superposition, then excited.
 
 
 def plot_raw_data_with_fit(
@@ -44,12 +21,12 @@ def plot_raw_data_with_fit(
     fits: Optional[xr.Dataset] = None,
 ) -> Figure:
     """
-    Plot All-XY I and Q (or state) vs sequence index with expected ideal values.
+    Plot All-XY state (or IQ amplitude) vs sequence index with expected ideal values.
 
     Parameters
     ----------
     ds : xr.Dataset
-        Dataset with coord sequence_index and per-qubit I/Q or state variables.
+        Dataset with coord sequence_index and per-qubit state or I/Q (IQ_abs) variables.
     qubits : list of AnyTransmon
         Qubits to plot.
     fits : xr.Dataset, optional
@@ -76,50 +53,25 @@ def _plot_one_qubit_all_xy(
     ds: xr.Dataset,
     qubit_info: dict,
 ) -> None:
-    """Plot I and Q vs sequence index for one qubit, with expected-value reference."""
+    """Plot state or IQ amplitude vs sequence index for one qubit, with expected reference."""
     n_seq = len(ALL_XY_LABELS)
     x = np.arange(n_seq)
-
     qubit_name = qubit_info["qubit"]
 
-    # Determine the numeric index of this qubit in the dataset's qubit coordinate,
-    # used as a fallback for numbered variable names (state1, state2, ...).
-    if "qubit" in ds.coords:
-        qubit_names = list(ds.coords["qubit"].values)
-        q_idx = qubit_names.index(qubit_name) if qubit_name in qubit_names else 0
+    if "state" in ds.data_vars:
+        y = ds["state"].sel(qubit=qubit_name).values
+        ylabel = "State population"
+        label = "State"
     else:
-        q_idx = 0
+        y = ds["IQ_abs"].sel(qubit=qubit_name).values
+        ylabel = "IQ amplitude [V]"
+        label = "IQ_abs"
 
-    state_var = f"state{q_idx + 1}"
-    i_var = f"I{q_idx + 1}"
-    q_var = f"Q{q_idx + 1}"
-
-    if "state" in ds.data_vars and "qubit" in ds["state"].dims:
-        state = ds["state"].sel(qubit=qubit_name).values
-        ax.plot(x, state, "bo-", label="State")
-        vmin, vmean, vmax = float(np.min(state)), float(np.mean(state)), float(np.max(state))
-        expected = [vmin] * N_GROUND + [vmean] * N_SUPERPOSITION + [vmax] * N_EXCITED
-        ax.plot(x, expected, "r-", label="Expected", alpha=0.8)
-        ax.set_ylabel("State population")
-    elif state_var in ds.data_vars:
-        state = ds[state_var].values
-        ax.plot(x, state, "bo-", label="State")
-        vmin, vmean, vmax = float(np.min(state)), float(np.mean(state)), float(np.max(state))
-        expected = [vmin] * N_GROUND + [vmean] * N_SUPERPOSITION + [vmax] * N_EXCITED
-        ax.plot(x, expected, "r-", label="Expected", alpha=0.8)
-        ax.set_ylabel("State population")
-    else:
-        I_vals = ds[i_var].values
-        Q_vals = ds[q_var].values
-        ax.plot(x, I_vals, "bo-", label="I")
-        ax.plot(x, Q_vals, "go-", label="Q")
-        imin, imean, imax = float(np.min(I_vals)), float(np.mean(I_vals)), float(np.max(I_vals))
-        qmin, qmean, qmax = float(np.min(Q_vals)), float(np.mean(Q_vals)), float(np.max(Q_vals))
-        expected_I = [imax] * N_GROUND + [imean] * N_SUPERPOSITION + [imin] * N_EXCITED
-        expected_Q = [qmax] * N_GROUND + [qmean] * N_SUPERPOSITION + [qmin] * N_EXCITED
-        ax.plot(x, expected_I, "r-", label="Expected I", alpha=0.8)
-        ax.plot(x, expected_Q, "m-", label="Expected Q", alpha=0.8)
-        ax.set_ylabel("Quadrature [a.u.]")
+    ax.plot(x, y, "bo-", label=label)
+    vmin, vmean, vmax = float(np.min(y)), float(np.mean(y)), float(np.max(y))
+    expected = [vmin] * N_GROUND + [vmean] * N_SUPERPOSITION + [vmax] * N_EXCITED
+    ax.plot(x, expected, "r-", label="Expected", alpha=0.8)
+    ax.set_ylabel(ylabel)
 
     ax.set_xlabel("Sequence")
     ax.set_xticks(x)
